lambda/src/requirements.txt
numpy-stl==2.17.1
numpy==1.24.3
